fix: type check
